requirements.txt
faiss-cpu>=1.7.4
numpy==1.26.4
requests==2.32.3
tqdm==4.67.1
yfinance==0.2.54
pandas==2.2.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
feedparser==6.0.11
beautifulsoup4==4.13.3
readability-lxml==0.8.1
lxml==5.3.2
fake-useragent==2.0.3
nltk==3.9.1
sentence-transformers==4.1.0
rank-bm25==0.2.2
python-dotenv==1.0.1
slowapi==0.1.9
APScheduler==3.10.4
httpx>=0.27.0
